requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.4

# Web Scraping & Browser Automation
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0

# Web UI
jinja2==3.1.2
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
thefuzz==0.20.0
python-Levenshtein==0.23.0
psutil==5.9.6
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
coverage==7.3.2
responses==0.24.1

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Development Tools
pre-commit==3.6.0
locust==2.17.0 